fix: add file to hold component constants
